+++ plugins/ops/apps-visualizer/test/integration/__init__.py
@@ -1,7 +1,4 @@
 from hopeit.app.config import AppConfig
-from hopeit.server.version import APPS_API_VERSION
-
-APP_VERSION = APPS_API_VERSION.replace('.', 'x')
 
 
 class MockAppEngine: